MTRAG-SemEval/mt_rag_solver/requirements.txt
transformers>=4.30.0
sentence-transformers>=2.2.2
faiss-cpu
rank_bm25
pyyaml
tqdm
torch>=1.10
